+++ packages/effects/plugins/src/echarts/index.ts
@@ -1,3 +1,4 @@
 export * from './echarts';
+export * from './types';
 export { default as EchartsUI } from './echarts-ui.vue';
 export * from './use-echarts';
